% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getmw}
\alias{getmw}
\title{Get the metabolomics workbench data}
\usage{
getmw(ID)
}
\arguments{
\item{ID}{study ID of metabolomics workbench}
}
\value{
list with details of checked study
}
\description{
Get the metabolomics workbench data
}
