% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getmwfactor}
\alias{getmwfactor}
\title{Convert metabolomics workbench factor into dataframe.}
\usage{
getmwfactor(ID)
}
\arguments{
\item{ID}{study ID of metabolomics workbench}
}
\value{
dataframe
}
\description{
Convert metabolomics workbench factor into dataframe.
}
