% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getmwlist}
\alias{getmwlist}
\title{Convert metabolomics workbench data into mzrt list object}
\usage{
getmwlist(ID)
}
\arguments{
\item{ID}{study ID of metabolomics workbench}
}
\value{
list
}
\description{
Convert metabolomics workbench data into mzrt list object
}
