% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getpmddf2}
\alias{getpmddf2}
\title{Get pmd details for specific reaction after the removal of isotopouge.}
\usage{
getpmddf2(dt, group = NULL, pmd = NULL, digits = 2, mdrange = c(0.25, 0.9))
}
\arguments{
\item{dt}{data.table of tims-MSI with first column as pixel ID and other columns for peaks with mz_ccs.}

\item{group}{mass to charge ratio group from either retention time or mass spectrometry imaging segmentation.}

\item{pmd}{a specific paired mass distance or a vector of pmds}

\item{digits}{mass or mass to charge ratio accuracy for pmd, default 2.}

\item{mdrange}{mass defect range to ignore. Default c(0.25,0.9) to retain the possible reaction related paired mass.}
}
\value{
dataframe with paired peaks for specific pmd or pmds. When group is provided, a column named net will be generated to show if certain pmd will be local(within the same group) or global(across the groups)
}
\description{
Get pmd details for specific reaction after the removal of isotopouge.
}
