% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb5}
\alias{hmdb5}
\title{HMDB MS2 spectra database}
\format{
HMDB 5.0 MS/MS spectra object for MS2 annotation. Download from https://github.com/jorainer/MetaboAnnotationTutorials/releases
}
\usage{
data(hmdb5)
}
\description{
HMDB MS2 spectra database
}
\keyword{datasets}
