% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdbcms}
\alias{hmdbcms}
\title{A list containing HMDB GC-EI-MS spectra database}
\format{
A list with compounds from hmdb for GC-MS simulation
}
\usage{
hmdbcms
}
\description{
A list containing HMDB GC-EI-MS spectra database
}
\keyword{datasets}
