% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hr}
\alias{hr}
\title{A data frame with compounds from hmdb and refmet}
\format{
A data frame with compounds from hmdb and refmet including name, InChIKey, chemical_formula, data source and exact mass
\describe{
  \item{name}{compounds name}
  \item{InChIKey}{InChIKey}
  \item{chemical_formula}{chemical formula}
  \item{dt}{database sources}
  \item{mass}{exact mass}
  }
}
\usage{
data(hr)
}
\description{
A data frame with compounds from hmdb and refmet
}
\keyword{datasets}
