% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monams1}
\alias{monams1}
\title{A list containing MoNA LC-MS spectra database}
\format{
A list with compounds from MoNA LC-MS spectra database for GC-MS simulation
}
\usage{
monams1
}
\description{
A list containing MoNA LC-MS spectra database
}
\keyword{datasets}
