% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mzrtn}
\alias{mzrtn}
\title{Demo mzrt list object for negative mode}
\format{
mzrt list object
}
\usage{
data(mzrtn)
}
\description{
Demo mzrt list object for negative mode
}
\keyword{datasets}
