% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paran}
\alias{paran}
\title{Demo ipo parameters object for negative mode}
\format{
ipo parameters
}
\usage{
paran
}
\description{
Demo ipo parameters object for negative mode
}
\keyword{datasets}
