% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srmnegeic}
\alias{srmnegeic}
\title{Demo xcmseic object for negative mode}
\format{
xcmseic object
}
\usage{
data(srmnegeic)
}
\description{
Demo xcmseic object for negative mode
}
\keyword{datasets}
