% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srmnegxset}
\alias{srmnegxset}
\title{Demo xcmsset object for negative mode}
\format{
xcmsset object
}
\usage{
data(srmnegxset)
}
\description{
Demo xcmsset object for negative mode
}
\keyword{datasets}
